post-service/requirements.txt
fastapi==0.112.4 
uvicorn==0.32.0 
pymongo==4.10.1 
python-dotenv==1.0.1 
httpx==0.27.2
python-multipart==0.0.6